LayoutTests/fast/frames/seamless/seamless-float-expected.txt

Test that floated seamless iframes 'shrink-wrap' their contents, as floated divs would.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(iframe1).width is "150px"
FAIL window.getComputedStyle(iframe1).height should be 50px. Was 104px.
PASS window.getComputedStyle(iframe2).width is "100px"
FAIL window.getComputedStyle(iframe2).height should be 100px. Was 104px.
